# server/openapi_spec.yaml
openapi: 3.0.0
info:
  title: Astronova Backend API
  description: Comprehensive astrology API with AI-powered insights, birth charts, compatibility matching, and personalized reports
  version: 1.0.0
  contact:
    name: Astronova Support
    email: admin@100xai.engineering
servers:
  - url: https://astronova-ghcr.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: "Admin API token. Production admin endpoints also accept Authorization: Bearer <ADMIN_API_TOKEN>."
  schemas:
...
    BirthData:
      type: object
      required:
        - date
        - time
        - timezone
        - latitude
        - longitude
      properties:
        date:
          type: string
          format: date
          example: "1990-05-15"
        time:
          type: string
          format: time
          example: "14:30"
        timezone:
          type: string
          example: "America/New_York"
        latitude:
          type: number
          format: float
          example: 40.7128
        longitude:
          type: number
          format: float
          example: -74.0060
...
    ChartPosition:
      type: object
      properties:
        sign:
          type: string
        degree:
          type: number
    SynastryAspect:
      type: object
      properties:
        planet1:
          type: string
        planet2:
          type: string
        aspect:
          type: string
        orb:
          type: number
...
paths:
...
  /api/v1/chart/generate:
    post:
      summary: Generate birth chart
      tags:
        - Charts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chartType
                - systems
                - birthData
              properties:
                chartType:
                  type: string
                  enum: [natal, transit, progressed]
                systems:
                  type: array
                  items:
                    type: string
                    enum: [western, vedic, chinese]
                birthData:
                  $ref: '#/components/schemas/BirthData'
      responses:
        '200':
          description: Generated chart data
          content:
            application/json:
              schema:
                type: object
                properties:
                  chartId:
                    type: string
                  type:
                    type: string
                  charts:
                    type: object
                    properties:
                      western:
                        type: object
                        properties:
                          svg:
                            type: string
                            description: Base64 encoded SVG
                          positions:
                            type: object
                            additionalProperties:
                              $ref: '#/components/schemas/ChartPosition'
                      vedic:
                        type: object
                        properties:
                          svg:
                            type: string
                          positions:
                            type: object
                      chinese:
                        type: object
                        properties:
                          svg:
                            type: string
                          animal:
                            type: string
                          element:
                            type: string
...
  /api/v1/chart/aspects:
    get:
      summary: Get aspects by date (UTC)
      tags:
        - Charts
      parameters:
        - in: query
          name: date
          schema:
            type: string
            example: "2025-01-01"
          required: true
      responses:
        '200':
          description: Aspect list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SynastryAspect'
    post:
      summary: Compute aspects from birth data
      tags:
        - Charts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                birthData:
                  $ref: '#/components/schemas/BirthData'
      responses:
        '200':
          description: Aspect list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SynastryAspect'
...